Create isPalindrome.py
The integer x is converted to a string using the str() function and stored in the variable str_x.

The [::-1] slicing notation is used to create a reversed version of the string str_x.

The original string str_x is compared with its reversed version using the == operator.

The result of the comparison is returned as the output of the isPalindrome method. If the string str_x is equal to its reverse, the function will return True, indicating that x is a palindrome. Otherwise, it will return False.