Create buddyStrings.py
We start by checking if the lengths of s and goal are different. If they are different, it's not possible to swap characters to make them equal. In this case, we immediately return False.

Next, we check if s and goal are already identical. If they are, we need to check if there are duplicate characters in s. If there are duplicates, we can swap two occurrences of the same character to get equal strings. We use the set() function to remove duplicate characters and compare the lengths. If the length of the set is less than the length of s, it means there are duplicates and we return True. Otherwise, we return False.

If s and goal are not identical, we need to find the indices where they differ. We iterate over the indices of s using a for loop and check if the characters at those indices are different between s and goal. If they are different, we append the index to the indices list.

We then check if the length of indices is exactly 2. If there are more or fewer than two differing indices, it's not possible to swap characters to make s and goal equal. In this case, we return False.

Finally, if there are exactly two differing indices, we retrieve the indices from the indices list and assign them to variables i and j. We check if swapping the characters at these indices in s and goal will result in equal strings. If s[i] is equal to goal[j] and s[j] is equal to goal[i], we return True. Otherwise, we return False.

By following these steps, we determine if it's possible to swap two letters in s to make it equal to goal.